refactor : 파일 컨셉 안맞는 메소드 위치 변경
